requirements.txt
streamlit>=1.33
pandas
pdfplumber
gspread
python-dotenv
openai>=1.40.0
google-generativeai>=0.7.2

